requirements.txt
fastapi~=0.101.1
motor~=3.2.0
paho-mqtt~=1.6.1
pymongo~=4.4.1
python-dotenv~=1.0.0
requests~=2.31.0
uvicorn~=0.23.2